Development

+++ identity_service/identity_service/Controllers/IdentityController.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿    using Microsoft.AspNetCore.Mvc;
 using Models.Auth;
 using Models.Exceptions;
 using Models.Identities;
 using Models.Services_Interfaces;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace identity_service.Controllers
 {
@@ -25,9 +26,14 @@ namespace identity_service.Controllers
                 Identity newIdentity = new Identity()
                 {
                     Email = req.Email.ToLower(),
-                    Password = req.Password,
+                    HashedPassword = req.Password,
+                    Birthdate = req.Birthdate,
                     RegisterMethod = "Website",
-                    Role = "User"
+                    Role = "User",
+                    UserLoginAttempt = new UserLoginAttempt()
+                    {
+                        AttemptsCount = 0
+                    }
                 };
 
 
@@ -51,6 +57,36 @@ namespace identity_service.Controllers
         {
             try
             {
+
+                identityService.DeleteIdentity(email.ToLower());
+                return NoContent();
+            }
+            catch (AssetAlreadyExistException)
+            {
+                return BadRequest("The email is already is been used.");
+            }
+            catch (Exception)
+            {
+                return StatusCode(500);
+            }
+        }
+
+
+        [HttpDelete("me")]
+        public IActionResult Delete()
+        {
+            try
+            {
+                string authorizationHeader = Request.Headers["Authorization"];
+                if (string.IsNullOrEmpty(authorizationHeader) || !authorizationHeader.StartsWith("Bearer "))
+                {
+                    return BadRequest("Invalid or missing Authorization header");
+                }
+                string token = authorizationHeader.Substring("Bearer ".Length).Trim();
+
+                var tokenHandler = new JwtSecurityTokenHandler();
+                var decodedToken = tokenHandler.ReadJwtToken(token);
+                var email = decodedToken.Claims.FirstOrDefault(c => c.Type == "Email")?.Value;
 
                 identityService.DeleteIdentity(email.ToLower());
                 return NoContent();
